Moved shield fetching to RouteManeuverViewController
RouteManeuverViewController is now responsible for fetching route shield images it needs to display. If the step has no code, try displaying a shield based on its destination code or, failing that, its destination.
