+++ MapboxNavigation/RouteManeuverViewController.swift
@@ -1,7 +1,7 @@
 import UIKit
 import MapboxDirections
 import MapboxCoreNavigation
-
+import SDWebImage
 
 class RouteManeuverViewController: UIViewController {
     @IBOutlet var separatorViews: [SeparatorView]!
@@ -19,6 +19,7 @@ class RouteManeuverViewController: UIViewController {
     weak var step: RouteStep! {
         didSet {
             if isViewLoaded {
+                roadCode = step.codes?.first ?? step.destinationCodes?.first ?? step.destinations?.first
                 updateStreetNameForStep()
             }
         }
@@ -42,12 +43,35 @@ class RouteManeuverViewController: UIViewController {
         return distance != nil ? 1 : 2
     }
     
+    var roadCode: String? {
+        didSet {
+            guard roadCode != oldValue, let components = roadCode?.components(separatedBy: " ") else {
+                return
+            }
+            
+            if components.count > 1 {
+                shieldAPIDataTask = dataTaskForShieldImage(network: components[0], number: components[1], height: 32 * UIScreen.main.scale) { [weak self] (image) in
+                    self?.shieldImage = image
+                }
+                shieldAPIDataTask?.resume()
+                if shieldAPIDataTask == nil {
+                    shieldImage = nil
+                }
+            } else {
+                shieldImage = nil
+            }
+        }
+    }
     var shieldImage: UIImage? {
         didSet {
             shieldImageView.image = shieldImage
             updateStreetNameForStep()
         }
     }
+    
+    var shieldAPIDataTask: URLSessionDataTask?
+    var shieldImageDownloadToken: SDWebImageDownloadToken?
+    let webImageManager = SDWebImageManager.shared()
     
     var availableStreetLabelBounds: CGRect {
         return CGRect(origin: .zero, size: maximumAvailableStreetLabelSize)
@@ -80,6 +104,7 @@ class RouteManeuverViewController: UIViewController {
     
     deinit {
         suspendNotifications()
+        webImageManager.cancelAll()
     }
     
     func resumeNotifications() {
@@ -105,6 +130,45 @@ class RouteManeuverViewController: UIViewController {
         }
         
         turnArrowView.step = routeProgress.currentLegProgress.upComingStep
+    }
+
+    func dataTaskForShieldImage(network: String, number: String, height: CGFloat, completion: @escaping (UIImage?) -> Void) -> URLSessionDataTask? {
+        guard let imageNamePattern = ShieldImageNamesByPrefix[network] else {
+            return nil
+        }
+
+        let imageName = imageNamePattern.replacingOccurrences(of: " ", with: "_").replacingOccurrences(of: "{ref}", with: number)
+        let apiURL = URL(string: "https://commons.wikimedia.org/w/api.php?action=query&format=json&maxage=86400&prop=imageinfo&titles=File%3A\(imageName)&iiprop=url%7Csize&iiurlheight=\(Int(round(height)))")!
+
+        shieldAPIDataTask?.cancel()
+        return URLSession.shared.dataTask(with: apiURL) { [weak self] (data, response, error) in
+            var json: [String: Any] = [:]
+            if let data = data, response?.mimeType == "application/json" {
+                do {
+                    json = try JSONSerialization.jsonObject(with: data, options: []) as! [String: Any]
+                } catch {
+                    assert(false, "Invalid data")
+                }
+            }
+
+            guard data != nil && error == nil else {
+                return
+            }
+
+            guard let query = json["query"] as? [String: Any],
+                let pages = query["pages"] as? [String: Any], let page = pages.first?.1 as? [String: Any],
+                let imageInfos = page["imageinfo"] as? [[String: Any]], let imageInfo = imageInfos.first,
+                let thumbURLString = imageInfo["thumburl"] as? String, let thumbURL = URL(string: thumbURLString) else {
+                    return
+            }
+
+            if thumbURL != self?.shieldImageDownloadToken?.url {
+                self?.webImageManager.imageDownloader?.cancel(self?.shieldImageDownloadToken)
+            }
+            self?.shieldImageDownloadToken = self?.webImageManager.imageDownloader?.downloadImage(with: thumbURL, options: .scaleDownLargeImages, progress: nil) { (image, data, error, isFinished) in
+                completion(image)
+            }
+        }
     }
     
     func updateStreetNameForStep() {
